Revert "Revert "Cleanups""
