+++ src/main/java/org/networkcalculus/dnc/AlgDncBackend.java
@@ -29,6 +29,7 @@ import org.networkcalculus.dnc.algebra.MinPlus;
 import org.networkcalculus.dnc.bounds.BoundingCurves;
 import org.networkcalculus.dnc.bounds.Bounds;
 import org.networkcalculus.dnc.curves.Curve;
+import org.networkcalculus.dnc.curves.CurveUtils;
 import org.networkcalculus.dnc.curves.LinearSegment;
 
 public interface AlgDncBackend {
@@ -39,6 +40,8 @@ public interface AlgDncBackend {
 	Bounds getBounds();
 	
 	Curve getCurveFactory();
+	
+	CurveUtils getCurveUtils();
 	
 	LinearSegment.Builder getLinearSegmentFactory();
