+++ src/main/java/org/networkcalculus/dnc/AlgDncBackend_DNC_Affine.java
@@ -32,9 +32,11 @@ import org.networkcalculus.dnc.bounds.Bounds;
 import org.networkcalculus.dnc.bounds.disco.BoundingCurves_Disco_ConPwAffine;
 import org.networkcalculus.dnc.bounds.disco.Bounds_Disco_PwAffine;
 import org.networkcalculus.dnc.curves.Curve;
+import org.networkcalculus.dnc.curves.CurveUtils;
 import org.networkcalculus.dnc.curves.LinearSegment;
 import org.networkcalculus.dnc.curves.disco.LinearSegment_Disco;
 import org.networkcalculus.dnc.curves.disco.affine.Curve_Disco_Affine;
+import org.networkcalculus.dnc.curves.disco.pw_affine.CurveUtils_Disco_PwAffine;
 
 public enum AlgDncBackend_DNC_Affine implements AlgDncBackend {
 	DISCO_AFFINE;
@@ -57,6 +59,11 @@ public enum AlgDncBackend_DNC_Affine implements AlgDncBackend {
 	@Override
 	public Curve getCurveFactory() {
 		return Curve_Disco_Affine.getFactory();
+	}
+
+	@Override
+	public CurveUtils getCurveUtils() {
+		return CurveUtils_Disco_PwAffine.getInstance();
 	}
 
 	@Override
