+++ src/main/java/org/networkcalculus/dnc/Calculator.java
@@ -28,6 +28,7 @@ package org.networkcalculus.dnc;
 
 import org.networkcalculus.dnc.algebra.MinPlus;
 import org.networkcalculus.dnc.curves.Curve;
+import org.networkcalculus.dnc.curves.CurveUtils;
 import org.networkcalculus.num.NumBackend;
 
 /**
@@ -68,6 +69,10 @@ public final class Calculator {
 	
 	public Curve getCurveFactory() {
 		return DNC_BACKEND.getCurveFactory();
+	}
+	
+	public CurveUtils getCurveUtils() {
+		return DNC_BACKEND.getCurveUtils();
 	}
 	
 	public MinPlus getMinPlus() {
