+++ src/main/java/org/networkcalculus/dnc/algebra/MinPlus.java
@@ -69,9 +69,6 @@ public interface MinPlus {
 
 	ArrivalCurve deconvolve(ArrivalCurve arrival_curve, ServiceCurve service_curve) throws Exception;
 
-	ArrivalCurve deconvolve(ArrivalCurve arrival_curve, ServiceCurve service_curve, boolean tb_rl_optimized)
-			throws Exception;
-
 	Set<ArrivalCurve> deconvolve_almostConcCs_SCs(Set<Curve> curves, Set<ServiceCurve> service_curves)
 			throws Exception;
 }
