+++ src/main/java/org/networkcalculus/dnc/feedforward/arrivalbounds/AggregatePboo_Concatenation.java
@@ -129,7 +129,7 @@ public class AggregatePboo_Concatenation extends AbstractArrivalBound implements
 			Set<ArrivalCurve> alphas_xxfcaller_s = new HashSet<ArrivalCurve>();
 			for (ArrivalCurve arrival_curve_path : alpha_xxfcaller_path) {
 				for (ArrivalCurve arrival_curve_offpath : alpha_xxfcaller_offpath) {
-					alphas_xxfcaller_s.add(Curve.add(arrival_curve_path, arrival_curve_offpath));
+					alphas_xxfcaller_s.add(Curve.getUtils().add(arrival_curve_path, arrival_curve_offpath));
 				}
 			}
