+++ src/main/java/org/networkcalculus/dnc/network/server_graph/Path.java
@@ -26,7 +26,6 @@
 
 package org.networkcalculus.dnc.network.server_graph;
 
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedList;
